split out app.py in orchestrator
