+++ src/playground-orchestrator/app/app.py
@@ -4,191 +4,15 @@ import logging
 import signal
 import sys
 import uuid
-from typing import Optional
-from datetime import datetime
-from sqlmodel import Field, create_engine, Session, select
-from sqlmodel import SQLModel
 from kubernetes import client, config
-from kubernetes.client import (
-    V1ObjectMeta,
-    V1PodSpec, 
-    V1Container, 
-    V1Pod, 
-    V1EnvVar, 
-    V1Volume, 
-    V1VolumeMount, 
-    V1EmptyDirVolumeSource,
-)
-from .config import (
-    get_environment_image, 
-    get_playground_controller_image,
-    get_backend_socketio_endpoint,
-    get_jwt_signing_key,
-    get_termination_grace_period,
-    get_image_pull_secret
-)
+from .database import get_reserved_sessions, get_unreserved_sessions, assign_session, unassign_session, remove_session
+from .pod_creation import create_new_pod
 
 # Configure logging
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 
 # Load the Kubernetes configuration
 config.load_incluster_config()
-
-# Database model for the playground_sessions table
-class PlaygroundSession(SQLModel, table=True):
-    __tablename__ = "playground_sessions"
-    
-    id: uuid.UUID                       = Field(default_factory=uuid.uuid4, primary_key=True)
-    type: str                           = Field(nullable=False)
-    instance_hostname: Optional[str]    = Field(nullable=True)
-    created_at_utc: datetime            = Field(default_factory=datetime.utcnow)
-
-def get_db_session():
-    connection_string = os.getenv("POSTGRES_CONNECTION_STRING")
-    engine = create_engine(connection_string)
-    session = Session(engine)
-    return session
-
-def get_unreserved_sessions():
-    session = get_db_session()
-    statement = select(PlaygroundSession).where(PlaygroundSession.instance_hostname == None)
-    result = session.exec(statement).all()
-    session.close()
-    return result
-
-def remove_session(session_id: uuid.UUID):
-    session = get_db_session()
-    statement = select(PlaygroundSession).where(PlaygroundSession.id == session_id)
-    result = session.exec(statement).one()
-    session.delete(result)
-    session.commit()
-    session.close()
-
-def assign_session(session_id: uuid.UUID, instance_hostname: str):
-    """
-    Updates a session with the instance hostname of the pod it is assigned to
-
-    Args:
-        session_id (uuid.UUID): The session ID
-        instance_hostname (str): The hostname of the pod
-    """
-    
-    db_session = get_db_session()
-    statement = select(PlaygroundSession).where(PlaygroundSession.id == session_id)
-    result = db_session.exec(statement).one()
-    result.instance_hostname = instance_hostname
-    db_session.commit()
-
-def unassign_session(session_id: uuid.UUID):
-    session = get_db_session()
-    statement = select(PlaygroundSession).where(PlaygroundSession.id == session_id)
-    result = session.exec(statement).one()
-    result.instance_hostname = None
-    session.add(result)
-    session.commit()
-    session.close()
-
-def get_reserved_sessions():
-    session = get_db_session()
-    statement = select(PlaygroundSession).where(PlaygroundSession.instance_hostname != None)
-    result = session.exec(statement).all()
-    session.close()
-    return result
-
-def create_new_pod(session_id: str, instance_type: str):
-    """
-    Creates a new pod for a playground session
-
-    Args:
-        session_id (str): The session ID
-        instance_type (str): The flavor of the environment
-
-    Returns:
-        str: The name of the created pod
-    """
-    
-    v1 = client.CoreV1Api()
-    
-    # Define the pod
-    pod = V1Pod(
-        metadata=V1ObjectMeta(
-            name=f"playground-{session_id}",
-            labels={
-                "app": "playground",
-                "playground-session-id": session_id
-            }
-        ),
-        spec=V1PodSpec(
-            termination_grace_period_seconds=get_termination_grace_period(),
-            containers=[
-                V1Container(
-                    name="playground-controller", 
-                    image=get_playground_controller_image(),
-                    volume_mounts=[
-                        V1VolumeMount(
-                            name="playground-shared",
-                            mount_path="/userland"
-                        ),
-                        V1VolumeMount(
-                            name="playground-hidden",
-                            mount_path="/playground",
-                            read_only=True
-                        )
-                    ],
-                    env=[
-                        V1EnvVar(
-                            name="BACKEND_SOCKETIO_ENDPOINT",
-                            value=get_backend_socketio_endpoint()
-                        ),
-                        V1EnvVar(
-                            name="JWT_SIGNING_KEY",
-                            value=get_jwt_signing_key()
-                        ),
-                        V1EnvVar(
-                            name="SESSION_ID",
-                            value=session_id
-                        )
-                    ]
-                ),
-                V1Container(
-                    name=f"{instance_type}-environment",
-                    image=get_environment_image(instance_type),
-                    volume_mounts=[
-                        V1VolumeMount(
-                            name="playground-shared",
-                            mount_path="/userland"
-                        ),
-                        V1VolumeMount(
-                            name="playground-hidden",
-                            mount_path="/playground"
-                        )
-                    ]
-                )
-            ],
-            volumes=[
-                V1Volume(
-                    name="playground-shared",
-                    empty_dir=V1EmptyDirVolumeSource()
-                ),
-                V1Volume(
-                    name="playground-hidden",
-                    empty_dir=V1EmptyDirVolumeSource()
-                )
-            ]
-        )
-    )
-    
-    image_pull_secret = get_image_pull_secret()
-    
-    if image_pull_secret:
-        pod.spec.image_pull_secrets = [{"name": image_pull_secret}]
-    
-    # Create the pod
-    v1.create_namespaced_pod(namespace="eduvize", body=pod)
-    
-    logging.info("Created new pod")
-    
-    return pod.metadata.name
 
 def get_pods():
     """
